Check/Assert that an Object is a Valid ESSENCE Field ID

Check/Assert that an Object is a Valid ESSENCE Field ID

check_field(
  field,
  datasource,
  creds = es_creds_get(),
  check_datasource = TRUE,
  fmatch = TRUE
)

assert_field(
  field,
  datasource,
  creds = es_creds_get(),
  check_datasource = TRUE,
  fmatch = TRUE,
  add = NULL
)

Arguments

datasource
[character(1)]. A datasource name (see es_datasources() for choices)

creds
secret. User login credentials for ESSENCE.

fmatch
[logical(1)]
Use the set operations implemented in fmatch in package fastmatch. If fastmatch is not installed, this silently falls back to match. fmatch modifies y by reference: A hash table is added as attribute which is used in subsequent calls.

add
[AssertCollection]
Collection to store assertion messages. See AssertCollection.

Value

check_*() returns TRUE if successful or a error message string otherwise; assert_*() returns x (invisibly) if successful or an error otherwise